JSDom tests in separate file
