+++ packages/wrapper/jest.config.jsdom.js
@@ -2,6 +2,6 @@ const config = require('next-redux-wrapper-configs/jest.config');
 module.exports = Object.assign({}, config, {
     roots: ['<rootDir>/src'],
     testPathIgnorePatterns: config.testPathIgnorePatterns.concat(['./es6', './lib', './types']),
-    testRegex: '/src/.*\\.test\\.tsx?$',
-    testEnvironment: 'node',
+    testRegex: '/src/.*\\.spec\\.tsx?$',
+    testEnvironment: 'jsdom',
 });
